pyproject.toml
[tool.poetry]
name = "houseplants-server"
version = "0.1.0"
description = ""
authors = ["Dianing Yudono"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
graphene = "^2.1.9"
graphene-django = "^2.15.0"
django-cors-headers = "^3.10.0"
pytest-django = "^4.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
